Add `browser-watch` script (Hot-Module Replacement)
Add an NPM script that uses `webpack-dev-server` to provide a
HMR-enabled server on port 8080.  Note: While the HMR support is
awesome, in its current state, this workflow doesn’t include
`cordova.js`.  A future TODO would be to analyze webpack-cordova-plugin
(https://github.com/markmarijnissen/webpack-cordova-plugin) as a
solution for this.  I tried, but couldn’t get it to work.